💄 style/모바일 검색창 뒷 배경, 최근 검색기록 전체 삭제, 그리드로 배열 변경#12
