+++ src/app/layout.tsx
@@ -1,4 +1,4 @@
-import { Header } from '@/components/Header';
+import { Header } from '@/components/Header/Header';
 import './globals.css';
 import Footer from '@/components/Footer';
